Replace views with s3 hosted CSVs
- Remove view piplines and replace with s3 backed csvs
- Split csv pipelines out into daily, monthly, yearly files
- Adds a new daily pipeline that recreates all non-static csv files
- Make datasets db a config item
