+++ dataflow/config.py
@@ -45,3 +45,7 @@ MATCHING_SERVICE_HAWK_KEY = os.environ.get('MATCHING_SERVICE_HAWK_KEY')
 MATCHING_SERVICE_HAWK_ALGORITHM = os.environ.get(
     'MATCHING_SERVICE_HAWK_ALGORITHM', 'sha256'
 )
+
+DATA_WORKSPACE_S3_BUCKET = os.environ.get("DATA_WORKSPACE_S3_BUCKET")
+DATASETS_DB_NAME = os.environ.get("DATASETS_DB_NAME")
+
