+++ dataflow/dags/dataset_pipelines.py
@@ -1,5 +1,4 @@
 """A module that defines Airflow DAGS for dataset pipelines."""
-
 from datetime import datetime, timedelta
 
 import sqlalchemy as sa
@@ -19,10 +18,11 @@ from dataflow.operators.db_tables import (
 
 
 class BaseDatasetPipeline:
-    target_db = 'datasets_db'
+    target_db = config.DATASETS_DB_NAME
     start_date = datetime(2019, 11, 5)
     end_date = None
     schedule_interval = '@daily'
+    catchup = False
 
     @property
     def table(self):
@@ -38,7 +38,7 @@ class BaseDatasetPipeline:
     def get_dag(self):
         with DAG(
             self.__class__.__name__,
-            catchup=False,
+            catchup=self.catchup,
             default_args={
                 'owner': 'airflow',
                 'depends_on_past': False,
@@ -46,6 +46,7 @@ class BaseDatasetPipeline:
                 'email_on_retry': False,
                 'retries': 0,
                 'retry_delay': timedelta(minutes=5),
+                'catchup': self.catchup,
             },
             start_date=self.start_date,
             end_date=self.end_date,
@@ -95,6 +96,7 @@ class BaseDatasetPipeline:
                 provide_context=True,
                 trigger_rule="all_done",
                 op_args=[self.target_db, self.table],
+                op_kwargs={'cascade': True},
             )
 
         (
