+++ sample.env
@@ -7,7 +7,7 @@ AIRFLOW__CELERY__RESULT_BACKEND=db+postgresql://data-flow-db/airflow
 AIRFLOW__CELERY__WORKER_CONCURRENCY=4
 
 AIRFLOW_CONN_DATASETS_DB=postgresql+psycopg2://data-flow-db/datasets
-AIRFLOW_CONN_DEFAULT_S3= s3://s3_connection_uri@S3
+AIRFLOW_CONN_DEFAULT_S3=s3://s3_connection_uri@S3
 
 AIRFLOW__CORE__DAGS_ARE_PAUSED_AT_CREATION=True
 AIRFLOW__CORE__DAGS_FOLDER=/app/dataflow/dags
@@ -56,3 +56,7 @@ MATCHING_SERVICE_UPDATE=False
 MATCHING_SERVICE_HAWK_ID=dataflow
 MATCHING_SERVICE_HAWK_KEY=some-key
 MATCHING_SERVICE_HAWK_ALGORITHM=sha256
+
+AIRFLOW_CONN_DATA_WORKSPACE_S3=s3://s3_connection_uri@S3
+DATA_WORKSPACE_S3_BUCKET=testing
+DATASETS_DB_NAME=datasets_db
