+++ tests/test_dags.py
@@ -6,7 +6,12 @@ from dataflow.dags import (
     canary,
     dataset_pipelines,
     ons_pipelines,
-    view_pipelines,
+)
+from dataflow.dags.csv_pipelines import (
+    csv_pipelines_daily,
+    csv_pipelines_monthly,
+    csv_pipelines_yearly,
+    csv_refresh_pipeline,
 )
 
 
@@ -23,7 +28,15 @@ def test_canary_tweets():
 
 @pytest.mark.parametrize(
     "module",
-    [activity_stream_pipelines, dataset_pipelines, ons_pipelines, view_pipelines],
+    [
+        activity_stream_pipelines,
+        csv_pipelines_daily,
+        csv_pipelines_monthly,
+        csv_pipelines_yearly,
+        csv_refresh_pipeline,
+        dataset_pipelines,
+        ons_pipelines,
+    ],
 )
 def test_pipelines_dags(module):
     dag_vars = [value for name, value in vars(module).items() if name.endswith("__dag")]
